added tone mapping back
