+++ interface/src/ui/Snapshot.cpp
@@ -41,7 +41,7 @@
 #include "MainWindow.h"
 #include "Snapshot.h"
 #include "SnapshotUploader.h"
-#include "ToneMappingEffect.h"
+#include "ToneMapAndResampleTask.h"
 
 // filename format: hifi-snap-by-%username%-on-%date%_%time%_@-%location%.jpg
 // %1 <= username, %2 <= date and time, %3 <= current location
