+++ libraries/render-utils/src/BloomEffect.cpp
@@ -15,7 +15,7 @@
 #include <shaders/Shaders.h>
 
 #include <render/BlurTask.h>
-#include <render/ToneMapAndResampleTask.h>
+#include "ToneMapAndResampleTask.h"
 #include "render-utils/ShaderConstants.h"
 
 #define BLOOM_BLUR_LEVEL_COUNT  3
