+++ libraries/render-utils/src/render-utils/toneMapping.slp
@@ -1,1 +1,2 @@
 VERTEX gpu::vertex::DrawViewportQuadTransformTexcoord
+DEFINES mirrored:f
