+++ libraries/render-utils/src/toneMapping.slf
@@ -43,7 +43,11 @@ layout(location=0) in vec2 varTexCoord0;
 layout(location=0) out vec4 outFragColor;
         
 void main(void) {
-    vec4 fragColorRaw = texture(colorMap, varTexCoord0);
+    <@if HIFI_USE_MIRRORED@>
+        vec4 fragColorRaw = texture(colorMap, vec2(1.0 - varTexCoord0.x, varTexCoord0.y));
+    <@else@>
+        vec4 fragColorRaw = texture(colorMap, varTexCoord0);
+    <@endif@>
     vec3 fragColor = fragColorRaw.xyz;
 
     vec3 srcColor = fragColor * getTwoPowExposure();
